% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airportCode}
\alias{airportCode}
\title{Airport Codes2}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(airportCode)
}
\description{
Airport Codes2
}
\keyword{datasets}
