% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_airports.R
\docType{data}
\name{airportCodeFull}
\alias{airportCodeFull}
\title{Airport Data - full}
\format{A dataframe with 6435 observations and 9 variables}
\usage{
airportCodeFull
}
\description{
USA airport data from the RITA/Transtats database
}
\keyword{datasets}
