% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airportMaster}
\alias{airportMaster}
\title{Airport Codes}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(airportMaster)
}
\description{
Airport Codes
}
\keyword{datasets}
