% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootnet.R
\name{boot_network}
\alias{boot_network}
\title{Network bootstrapping}
\usage{
boot_network(g, n = 500, left_ci = 0.005, right_ci = 0.995)
}
\arguments{
\item{g}{iGraph graph.}

\item{n}{Number of bootstraps to run. (500 default)}

\item{left_ci}{Confidence interval left limit. (0.005 default)}

\item{right_ci}{Confidence interval left limit (0.995 default)}
}
\description{
Bootstraps a network and returns output containing three network statistics:
Average Path Length, Transitivity, Mean Betweenness.
}
\examples{
\dontrun{
boot_net(g, n = 500)

}

}
