% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carrier}
\alias{carrier}
\title{Load data}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(carrier)
}
\description{
Carrier codes
}
\keyword{datasets}
