% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_carriers.R
\docType{data}
\name{carriers}
\alias{carriers}
\title{Carrier data}
\format{A dataframe with 1738 observations and 2 variables}
\usage{
carriers
}
\description{
This data comes from the RITA/Transtats database
}
\keyword{datasets}
