% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNodes.R
\name{createNodes}
\alias{createNodes}
\title{Create Nodes}
\usage{
createNodes(y)
}
\arguments{
\item{y}{Data Frame}
}
\description{
Creates nodes for SKYNET's functions.
Despite being possible to use it individually, it's mainly meant to be used as a complimentary function.
}
