% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparityfilter.R
\name{dispfilter}
\alias{dispfilter}
\title{Disparity Filter}
\usage{
dispfilter(g, alpha = 0.003)
}
\arguments{
\item{g}{igraph object}

\item{alpha}{Alpha value. Default 0.003}
}
\description{
Uses the Serrano's disparity filter (\url{http://www.pnas.org/content/106/16/6483.full})
to extract the backbone of the network in "Extracting the multiscale backbone of complex weighted networks"
}
\examples{
\dontrun{
dispfilter(g, alpha = 0.003)
}
}
