% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_db1b.R
\name{download_db1b}
\alias{download_db1b}
\title{Download Data from DB1B files}
\usage{
download_db1b(y = NULL, q = NULL)
}
\arguments{
\item{y}{year to be imported}

\item{q}{quarter to be imported}
}
\description{
Downloads data from BTS/RITA/Transtats files
}
\details{
Coupon files are downloaded from \url{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=289}.
Ticket files are downloaded from  \url{https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=272}.

Note: The BTS often changes the way we can access these files. So please be warned that this is still an experimental feature.
}
\examples{
\dontrun{

download_db1b(2010, 1)

}
}
