% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ontime.R
\name{download_ontime}
\alias{download_ontime}
\title{Download On-Time}
\usage{
download_ontime(y, m, auto = TRUE)
}
\arguments{
\item{y}{year to be imported}

\item{m}{month to be imported}

\item{auto}{Automatically assigns object}
}
\description{
Download On-Time Performance Data directly from BTS/RITA/Transtats website raw data (prezipped file),
for SKYNET's import function.
}
\examples{
\dontrun{

import_ontime(skynet_example("Ontime.csv"))

}
}
