% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_t100.R
\name{download_t100}
\alias{download_t100}
\title{Download Data from T100 files}
\usage{
download_t100(y = NULL, type = NULL)
}
\arguments{
\item{y}{year to be imported}

\item{type}{"mkt" for Market, "seg" for Segment databases respectively}
}
\description{
Downloads data from BTS/RITA/Transtats and imports it into R
}
\details{
Note: The BTS often changes the way we can access these files. So please be warned that this is still an experimental feature.
}
\examples{
\dontrun{

download_t100(2010, "mkt")

}
}
