% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAirport.R
\name{find_airport}
\alias{find_airport}
\title{Find Airport function}
\usage{
find_airport(x)
}
\arguments{
\item{x}{airport IATA code or city name}
}
\description{
Searches for airport information based on its IATA code or city name
It will display multiple airports as it works with partial names.
}
\examples{
\dontrun{
find_airport("Atlanta")

find_airport("ATL")
}
}
