% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCarrier.R
\name{find_carrier}
\alias{find_carrier}
\title{Find Carrier function}
\usage{
find_carrier(x)
}
\arguments{
\item{x}{Carrier}
}
\description{
Searches for airport information based on its IATA code or city name
}
\examples{
\dontrun{
find_carrier("United")

find_carrier("UA")
}
}
