% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findairport.R
\name{findAirport}
\alias{findAirport}
\title{Find Airport function}
\usage{
findAirport(x)
}
\arguments{
\item{x}{airport IATA code or city name}
}
\description{
Searches for airport information based on its IATA code or city name
}
\examples{
\dontrun{
findAirport("Atlanta")

findAirport("ATL")
}
}
