% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCarrier.R
\name{findCarrier}
\alias{findCarrier}
\title{Find Carrier function}
\usage{
findCarrier(x)
}
\arguments{
\item{x}{Carrier}
}
\description{
Searches for airport information based on its IATA code or city name
}
\examples{
\dontrun{
findCarrier("United")

findCarrier("UA")
}
}
