% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerLaw.R
\name{fitPlaw}
\alias{fitPlaw}
\title{Power Law}
\usage{
fitPlaw(graph)
}
\arguments{
\item{graph}{iGraph object}
}
\description{
Plots power law fit
}
\examples{
\dontrun{
netDir <- make.netDir(OD_Sample)
fitPlaw(netDir$gDir)
}
}
