% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromto.R
\name{from_to_stats}
\alias{from_to_stats}
\title{From To function}
\usage{
from_to_stats(x, y, orig)
}
\arguments{
\item{x}{igraph object to query}

\item{y}{origin airport IATA code}

\item{orig}{"from" or "to" options}
}
\description{
Calculate edges weight from IATA Code
}
\examples{
\dontrun{
netDir <- make.netDir(OD_Sample)
from_to_stats(netDir$gDir, "JFK", orig = "from")

from_to_stats(netDir$gDir, "JFK", orig = "to")
}

}
