% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromto.R
\name{fromto.stat}
\alias{fromto.stat}
\title{From To function}
\usage{
fromto.stat(x, y, orig)
}
\arguments{
\item{x}{igraph object to query}

\item{y}{origin airport IATA code}

\item{orig}{"from" or "to" options}
}
\description{
Calculate edges weight from IATA Code
}
\examples{
\dontrun{
netDir <- make.netDir(OD_Sample)
fromto.stat(netDir$gDir, "JFK", orig = "from")

fromto.stat(netDir$gDir, "JFK", orig = "to")
}

}
