% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_db1b.R
\name{import_db1b}
\alias{import_db1b}
\title{Import Data from DB1B files}
\usage{
import_db1b(x, y, zip = FALSE)
}
\arguments{
\item{x}{First csv file to be imported, in case of DB1B database}

\item{y}{Second csv file to be imported.}

\item{zip}{Should equal TRUE if original file comes from the BTS prezipped option.}
}
\description{
Imports data from BTS/RITA/Transtats website
File order doesn't matter, but it is recommended to name the files using the following
syntax: \emph{"Coupon YearQuarter.csv", "Ticket YearQuarter.csv".}
Note: We do recommend sparklyr to be used for larger sets of data.
}
\examples{
\dontrun{

import_db1b(skynet_example("Coupon_2001Q1.csv"), skynet_example("Ticket_2001Q1.csv"))

}
}
