% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ontime.R
\name{import_ontime}
\alias{import_ontime}
\title{Import on-time Data}
\usage{
import_ontime(x, auto = TRUE)
}
\arguments{
\item{x}{On-time csv (from zipped file)}

\item{auto}{Automatically assigns object}
}
\description{
Imports on-time Data directly from BTS/RITA/Transtats website raw data (prezipped file),
for SKYNET's import function.
}
\details{
Files can be found here \url{https://www.transtats.bts.gov/Fields.asp?gnoyr_VQ=FGJ}.
More information on variables to select and type of files to use can be found \href{https://github.com/ropensci/skynet}{here}
}
\examples{
\dontrun{

import_ontime(skynet_example("Ontime_2011_1.csv"))

}
}
