% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_t100.R
\name{import_t100}
\alias{import_t100}
\title{Import T-100 Data}
\usage{
import_t100(x, nonsch = FALSE)
}
\arguments{
\item{x}{T-100 csv}

\item{nonsch}{Should equal TRUE to include non-scheduled flights}
}
\description{
Imports T-100 Data directly from BTS/RITA/Transtats website raw data (prezipped file),
for SKYNET's import function.
Please try to name original csv as e.g. "T100 2011Q1mkt.csv" or "T100 2011Q1seg.csv"
respectively
}
\examples{
\dontrun{

import_t100(skynet_example("T100_2011_mkt.csv"))

}
}
