% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMap.R
\name{make.gMap}
\alias{make.gMap}
\title{Net Map}
\usage{
make.gMap(x, pct = 60)
}
\arguments{
\item{x}{Data frame}

\item{pct}{percentage of edges to include}
}
\description{
Creates OD ggplot2 generated maps from make.net functions
Shows sample of 60% of flights
}
\examples{
\dontrun{
make.gMap(OD_2016Q1, pct = 10)
}
}
