% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMetro.R
\name{make.netMetro}
\alias{make.netMetro}
\title{Metro Network}
\usage{
make.netMetro(x = NULL, undirected = FALSE, merge = TRUE)
}
\arguments{
\item{merge}{When set to FALSE, it keeps parallel edges instead of collapsing them
and summing their weights.}

\item{disp}{Uses the Serrano's disparity filter (\url{http://www.pnas.org/content/106/16/6483.full})
to extract the backbone of the network.}

\item{alpha}{Argument for disparity filter.}

\item{cap}{Filters original data based on the edge weight.}

\item{pct}{Argument for cap filter. Value should be imput as percentage.}
}
\description{
Generates Metro Network with an iGraph object and a Data Frame.
Still to fix some details
}
\examples{
make.netMetro(OD_2016Q1)

# Apply Disparity Filter
make.netUnd(OD_2016Q1, disp = TRUE, alpha = 0.05)

# Apply Percentage Cap
make.netUnd(OD_2016Q1, cap = TRUE, pct = 20)
}
