% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodesTr.R
\name{make.nodesTr}
\alias{make.nodesTr}
\title{Transfer Nodes}
\usage{
make.nodesTr(x)
}
\arguments{
\item{x}{Data set}
}
\description{
Extracts Nodes used only by transfer passengers
}
\examples{
make.nodesTr(OD_2016Q1)

}
