% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_net_trip.R
\name{make_net_trip}
\alias{make_net_trip}
\title{Trip directed network}
\usage{
make_net_trip(x, carrier = FALSE)
}
\arguments{
\item{x}{Data frame}

\item{carrier}{Groups data per carrier and OD}
}
\description{
Generates Trip/Route based Directed Network with an iGraph \strong{gDir} object,
a Data Frame \strong{netDir} and a Data Frame
with Airport/Nodes statistics \strong{nodes}.
Returns type of trip:
OD = Origin/Destination pair,
OT = Origin/Transfer pair,
TT = Transfer/Transfer pair,
TD = Transfer/Destination pair
}
\examples{
\dontrun{
make_net_trip(OD_Sample)
}
}
