% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_metro.R
\docType{data}
\name{MetroFull}
\alias{MetroFull}
\title{Metro (Full) Data}
\format{A dataframe with 5802 observations and 5 variables}
\usage{
MetroFull
}
\description{
This data comes from the RITA/Transtats database
}
\keyword{datasets}
