% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_metro.R
\docType{data}
\name{MetroLookup}
\alias{MetroLookup}
\title{Metro Data}
\format{A dataframe with 5802 observations and 2 variables}
\usage{
MetroLookup
}
\description{
This data comes from the RITA/Transtats database
}
\keyword{datasets}
