% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNodes.R
\name{metroNodes}
\alias{metroNodes}
\title{Create Transfer nodes}
\usage{
metroNodes(y)
}
\arguments{
\item{y}{Data Frame}
}
\description{
#@param x Original Data Frame to extract nodes from
#@param y List to include it into
}
\details{
Create Metro Nodes
}
