% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{net_map}
\alias{net_map}
\title{Plot Skynet}
\usage{
net_map(x, pct = 60)
}
\arguments{
\item{x}{Skynet Object
(generated by make_net_dir,make_net_und or make_net_path)}

\item{pct}{percentage of edges to include}
}
\description{
Creates OD ggplot2 generated maps from make.net functions
Shows sample of 60% of flights
}
\examples{
\dontrun{
network <- make.netDir(OD_Sample)
net_map(network, pct = 10)
}
}
