% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netImport.R
\name{netImport}
\alias{netImport}
\title{Skynet network analysis for BTS Data
Creates networks from the BTS Airline Origin and Destination Survey (DB1B)}
\usage{
netImport(x, y)
}
\description{
Given the DB1BCoupon and DB1BTicket csv's exported
this package allows creating sociomatrixes and subsequent igraph graphs.
}
\examples{
NA
}
\references{
NA.
NA.
}
\author{
Filipe Teixeira
}
