% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{netMap}
\alias{netMap}
\title{netMap}
\usage{
netMap(x, pct = 60)
}
\arguments{
\item{x}{Data frame}

\item{pct}{percentage of edges to include}
}
\description{
Creates OD ggplot2 generated maps from make.net functions
Shows sample of 60% of flights
}
\examples{
\dontrun{
network <- make.netDir(OD_Sample)
netMap(network$netDir, pct = 10)
}
}
