% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeStats.R
\name{node_stats}
\alias{node_stats}
\title{Get node info}
\usage{
node_stats(x)
}
\arguments{
\item{x}{Data Frame to extract information from}
}
\description{
Creates node statistics
Generates Number of Passenger Arrivals, Departures and Transfers
}
\examples{
\dontrun{

node_stats(OD_Sample)

}

}
