% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeStats.R
\name{nodeStats}
\alias{nodeStats}
\title{Get node info}
\usage{
nodeStats(x)
}
\arguments{
\item{x}{Data Frame to extract information from}
}
\description{
Creates node statistics
Generates Number of Passenger Arrivals, Departures and Transfers
}
\examples{
\dontrun{

nodeStats(OD_2011Q1)

}

}
