% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skynet.R
\docType{package}
\name{skynet-package}
\alias{skynet}
\alias{skynet-package}
\title{skynet: Network analysis for BTS Data}
\description{
Creates networks from the BTS/Transtats data
}
\details{
Given the DB1BCoupon and DB1BTicket, or the T-100 csv's exported
this package allows creating sociomatrixes and subsequent igraph graphs.
}
\examples{
NA

}
\references{
NA
}
\author{
Filipe Teixeira
}
\keyword{#@import}
\keyword{#@importFrom}
\keyword{data.table}
\keyword{getBackboneNetwork}
\keyword{internal}
\keyword{semnet}
