% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skynet_example.R
\name{skynet_example}
\alias{skynet_example}
\title{Get path to skynet examples}
\usage{
skynet_example(path = NULL)
}
\arguments{
\item{path}{File name.}
}
\description{
To access csv examples from SKYNET
}
\examples{
\dontrun{
skynet_example()
skynet_example("Coupon 2001Q1.csv")
}
}
