% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.skynet}
\alias{summary.skynet}
\title{Displays a summary of a skynet object}
\usage{
\method{summary}{skynet}(object, ...)
}
\arguments{
\item{object}{skynet object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Displays a summary of a skynet object
}
\examples{
net <- make_net_dir(OD_Sample)
summary(net)
}
