% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNodes.R
\name{transferNodes}
\alias{transferNodes}
\title{Create Transfer nodes}
\usage{
transferNodes(x, y)
}
\arguments{
\item{x}{Original Data Frame to extract nodes from}

\item{y}{List to include it into}
}
\description{
Create Transfer nodes
}
