% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_airports.R
\docType{data}
\name{airportCode}
\alias{airportCode}
\title{Airport Data - clean}
\format{A dataframe with 6435 observations and 5 variables}
\usage{
airportCode
}
\description{
USA airport data from the RITA/Transtats database
}
\keyword{datasets}
