% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_airports.R
\docType{data}
\name{airportMaster}
\alias{airportMaster}
\title{Airport Data - master}
\format{A dataframe with 13555 observations and 28 variables}
\usage{
airportMaster
}
\description{
World airport data from the RITA/Transtats database
}
\keyword{datasets}
