% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRaw.R
\name{convertRaw}
\alias{convertRaw}
\title{Convert Raw}
\usage{
convertRaw(x, y)
}
\arguments{
\item{x}{First csv file to be processed}

\item{y}{Second csv file to be processed}
}
\description{
Converts T100 and DB1B files from BTS/RITA/Transtats website raw data (prezipped file), for SKYNET's import function.
}
\details{
File order doesn't matter, but it is recomended to name the files using the following
syntax: \emph{"Origin_and_Destination_Survey_DB1BCoupon_year_quarter.csv", "Origin_and_Destination_Survey_DB1BTicket_year_quarter.csv".}
Note: We do recommend sparklyr to be used for larger sets of data.
}
\examples{

# DB1B Database Files - Ticket and Coupon order doesn't matter
\dontrun{
netConvert("Origin_and_Destination_Survey_DB1BCoupon_2017_1.csv",
"Origin_and_Destination_Survey_DB1BTicket_2017_1.csv")
}
}
