% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_t100.R
\name{import_t100}
\alias{import_t100}
\title{Import T-100 Data}
\usage{
import_t100(x, nonsch = FALSE, auto = TRUE)
}
\arguments{
\item{x}{T-100 csv}

\item{nonsch}{Should equal TRUE to include non-scheduled flights}

\item{auto}{Automatically assigns object}
}
\description{
Imports T-100 Data directly from BTS/RITA/Transtats website raw data (prezipped file),
for SKYNET's import function.
}
\details{
Files can be found here \url{https://www.transtats.bts.gov/Tables.asp?DB_ID=111}.
More information on variables to select and type of files to use can be found \href{https://github.com/FilipeamTeixeira/skynet}{here}
}
\examples{
\dontrun{

import_t100(skynet_example("T100_2011_mkt.csv"))

}
}
