% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importtest.R
\name{importT100}
\alias{importT100}
\title{Import T-100 Data}
\usage{
importT100(x)
}
\arguments{
\item{x}{T100 csv}
}
\description{
#Don't forget to add information on segment or market (to test)
}
\examples{
\dontrun{
importT100(x)
}
}
