% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importT100.R
\name{importT100}
\alias{importT100}
\title{Import T-100 Data}
\usage{
importT100(x)
}
\arguments{
\item{x}{T100 csv}
}
\description{
Imports T-100 Data directly from zipped file
Please try to name original csv as e.g. "T100 2011Q1mkt.csv" or "T100 2011Q1seg.csv"
respectively
}
\examples{
NA

}
