% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netDir.R
\name{make.netDir}
\alias{make.netDir}
\title{Directed network}
\usage{
make.netDir(x, disp = FALSE, cap = FALSE, alpha = 0.003, pct = 10)
}
\arguments{
\item{x}{Data frame}

\item{disp}{Uses the Serrano's disparity filter (\url{http://www.pnas.org/content/106/16/6483.full})
to extract the backbone of the network.}

\item{cap}{Filters original data based on the edge weight.}

\item{alpha}{Argument for disparity filter.}

\item{pct}{Argument for cap filter. Value should be imput as percentage.}
}
\description{
Generates Directed Network with an iGraph object and a Data Frame.
}
\examples{
make.netDir(OD_2016Q1)

# Apply Disparity Filter
make.netDir(OD_2016Q1, disp = TRUE, alpha = 0.05)

# Apply Percentage Cap
make.netDir(OD_2016Q1, cap = TRUE, pct = 20)

}
