% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMetro.R
\name{make.netMetro}
\alias{make.netMetro}
\title{Metro Network}
\usage{
make.netMetro(x = NULL, undirected = FALSE, merge = TRUE)
}
\arguments{
\item{x}{data frame}

\item{undirected}{If true, creates undirected network}

\item{merge}{When set to FALSE, it keeps parallel edges instead of collapsing them
and summing their weights.}
}
\description{
Generates Metro Network with an iGraph object and a Data Frame.
Still to fix some details - IMPORTANT!!!!!!!!!!!!!!!
}
\examples{
\dontrun{
make.netMetro(OD_2016Q1)

# Creates undirected network with merged edges
make.netUnd(OD_2016Q1, undirected = TRUE, merge = TRUE)
}

}
