% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netPath.R
\name{make.Path}
\alias{make.Path}
\title{Path and OD Network}
\usage{
make.Path(x, leg = FALSE)
}
\arguments{
\item{x}{Data frame}

\item{leg}{Generates Leg Count Data frame, based on Path taken.
For example, all passengers doing the BOS-ATL-LAX path, are summed by Air Carrier.}
}
\description{
Generates an OD network and a Leg Count data frame(on demand)
}
\examples{
make.Path(OD_2016Q1)

# Generate Leg Count
make.Path(OD_2016Q1, leg = TRUE)

}
