% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MetroLookup}
\alias{MetroLookup}
\title{City Names for Metro merge}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(MetroLookup)
}
\description{
City Names for Metro merge
}
\keyword{datasets}
