% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netImport.R
\name{netImport}
\alias{netImport}
\title{Import Data}
\usage{
netImport(x, y)
}
\arguments{
\item{x}{First csv file to be imported, in case of DB1B database, or in case of using
the T-100 database, the only file to be included.}

\item{y}{Second csv file to be imported.}
}
\description{
Imports data from BTS/RITA/Transtats website
File order doesn't matter, but it is recomended to name the files using the following
syntax: \emph{"Coupon YearQuarter.csv", "Ticket YearQuarter.csv", "T100 Year".}
Note: We do recommend sparklyr to be used for larger sets of data.
}
\examples{

# DB1B Database Files - Ticket and Coupon order doesn't matter
netImport("Coupon 2016Q1.csv", "Ticket 2016Q1.csv")

# T-100
netImport("T100 2016.csv")

}
