% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netImport.R
\name{netImport}
\alias{netImport}
\title{Import Data}
\usage{
netImport(x, y, zip = FALSE, nonsch = FALSE)
}
\arguments{
\item{x}{First csv file to be imported, in case of DB1B database, or in case of using
the T-100 database, the only file to be included.}

\item{y}{Second csv file to be imported.}

\item{zip}{Should equal TRUE if original file comes from the BTS prezipped option.}

\item{nonsch}{Should equal TRUE to include non-scheduled flights}
}
\description{
Imports data from BTS/RITA/Transtats website
File order doesn't matter, but it is recommended to name the files using the following
syntax: \emph{"Coupon YearQuarter.csv", "Ticket YearQuarter.csv", "T100 Year".}
Note: We do recommend sparklyr to be used for larger sets of data.
}
\examples{
\dontrun{

netImport(skynet_example("Coupon_2001Q1.csv"), skynet_example("Ticket_2001Q1.csv"))

}
}
